+++ include/Small/Scripting/SFMLtoArk.hpp
@@ -1,0 +1,23 @@
+/**
+ * @file SFMLtoArk.hpp
+ * @author Alexandre Plateau (lexplt.dev@gmail.com)
+ * @brief Convert SFML values to ArkScript values
+ * @version 0.1
+ * @date 2020-04-19
+ * 
+ * @copyright Copyright (c) 2020
+ * 
+ */
+
+#ifndef sgl_small_scripting_sfmltoark
+#define sgl_small_scripting_sfmltoark
+
+#include <Ark/VM/Value.hpp>
+#include <SFML/Graphics.hpp>
+
+namespace sgl::Scripting
+{
+
+}
+
+#endif