+++ source/Scripting/SFMLtoArk.cpp
@@ -1,0 +1,4 @@
+#include <Small/Scripting/SFMLtoArk.hpp>
+
+namespace sgl::Scripting
+{}