+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - added image widget
 - adding `Small/Scripting/` and starting to develop solutions to ease the use of ArkScript for the library
 - adding file system utilities
+- adding a list of bindings from the small game library to ArkScript
 
 ### Changed
 - `thirdparty/` is now a cmake subdirectory